Move CoreDeployer to deploy pkg
