+++ typescript/deploy/src/index.ts
@@ -2,19 +2,21 @@ export { AbacusAppChecker, Ownable } from './check';
 export {
   CheckerViolation,
   EnvironmentConfig,
-  TransactionConfig,
+  TransactionConfig
 } from './config';
+export { AbacusCoreDeployer, CoreConfig, ValidatorManagerConfig } from './core/deploy';
 export { AbacusAppDeployer } from './deploy';
 export {
   ProxiedContract,
   ProxyViolationType,
-  UpgradeBeaconViolation,
+  UpgradeBeaconViolation
 } from './proxy';
 export {
   AbacusRouterChecker,
   AbacusRouterDeployer,
   Router,
-  RouterConfig,
+  RouterConfig
 } from './router';
 export * as utils from './utils';
 export { ContractVerifier, VerificationInput } from './verify';
+
