+++ typescript/infra/config/environments/dev/core.ts
@@ -1,8 +1,8 @@
+import { CoreConfig } from '@abacus-network/deploy';
 import { ChainMap } from '@abacus-network/sdk';
-
-import { CoreConfig } from '../../../src/core';
-
 import { DevNetworks } from './domains';
+
+
 
 export const core: ChainMap<DevNetworks, CoreConfig> = {
   alfajores: {
