+++ typescript/infra/config/environments/testnet/core.ts
@@ -1,8 +1,8 @@
+import { CoreConfig } from '@abacus-network/deploy';
 import { ChainMap } from '@abacus-network/sdk';
-
-import { CoreConfig } from '../../../src/core';
-
 import { TestnetNetworks } from './domains';
+
+
 
 export const core: ChainMap<TestnetNetworks, CoreConfig> = {
   alfajores: {
