+++ typescript/infra/hardhat.config.ts
@@ -1,26 +1,25 @@
-import '@nomiclabs/hardhat-etherscan';
-import '@nomiclabs/hardhat-waffle';
-import { task } from 'hardhat/config';
-import { HardhatRuntimeEnvironment } from 'hardhat/types';
-
 import { BadRandomRecipient__factory } from '@abacus-network/core';
 import { utils as deployUtils } from '@abacus-network/deploy';
 import {
   AbacusCore,
   ChainName,
-  ChainNameToDomainId,
+  ChainNameToDomainId
 } from '@abacus-network/sdk';
 import { utils } from '@abacus-network/utils';
-
+import '@nomiclabs/hardhat-etherscan';
+import '@nomiclabs/hardhat-waffle';
+import { task } from 'hardhat/config';
+import { HardhatRuntimeEnvironment } from 'hardhat/types';
 import {
   getCoreContractsSdkFilepath,
   getCoreEnvironmentConfig,
   getCoreRustDirectory,
-  getCoreVerificationDirectory,
+  getCoreVerificationDirectory
 } from './scripts/utils';
-import { AbacusCoreDeployer } from './src/core';
+import { AbacusCoreInfraDeployer } from './src/core/deploy';
 import { sleep } from './src/utils/utils';
 import { AbacusContractVerifier } from './src/verify';
+
 
 const domainSummary = async <Networks extends ChainName>(
   core: AbacusCore<Networks>,
@@ -78,7 +77,7 @@ task('abacus', 'Deploys abacus on top of an already running Hardhat Network')
       signer,
     );
 
-    const deployer = new AbacusCoreDeployer(multiProvider, config.core);
+    const deployer = new AbacusCoreInfraDeployer(multiProvider, config.core);
     const addresses = await deployer.deploy();
 
     // Write configs
