+++ typescript/infra/src/core/index.ts
@@ -1,4 +1,3 @@
-export { AbacusCoreDeployer } from './deploy';
 export { AbacusCoreChecker, CoreViolationType } from './check';
 export { AbacusCoreGovernor } from './govern';
-export { CoreConfig } from './types';
+
